LayoutTests/platform/mac-bigsur/imported/w3c/web-platform-tests/content-security-policy/style-src/style-src-injected-stylesheet-allowed.sub-expected.txt

Blocked access to external URL http://www1.localhost:8800/content-security-policy/style-src/resources/style-src.css

FAIL Programatically injected stylesheet should load assert_equals: expected "2px" but got "0px"
Lorem ipsum
